sixpc_reservation/requirements.txt
selenium==4.19.0
python-dotenv
schedule
requests